requirements.txt
pytest==8.4.2
selenium==4.35.0
webdriver-manager==4.0.2
requests==2.32.5
SQLAlchemy==1.4.50
python-dotenv==1.1.1
psycopg2-binary==2.9.10
allure-pytest==2.15.0
allure-python-commons==2.15.0
pytest-xdist==3.8.0